+++ CHANGELOG.md
@@ -29,6 +29,9 @@
     - Highly specific spawn configuration like this option provided is better suited for other mods to handle, not AVP.
 
 ## 🐞 Fixes
+- Fixed nether resin nodes having incorrect block properties.
+- Fixed irradiated resin, irradiated resin nodes, irradiated resin veins and irradiated resin webs having incorrect block properties.
+- Fixed resin, resin nodes, resin veins and resin webs having incorrect colors on maps.
 - Fixed variant aliens being unable to replace enemy variant resin.
 - Fixed variant aliens not creating hives near enemy variant hives.
 - Fixed variant aliens spawning on resin near enemy variant hives without having a nearby hive of their own to spawn in.